update di arch with stone 1.0.7
