+++ app_mobile/src/main/kotlin/com/github/klee0kai/thekey/app/di/AppComponent.kt
@@ -4,21 +4,11 @@ import com.github.klee0kai.stone.KotlinWrappersStone
 import com.github.klee0kai.stone.Stone
 import com.github.klee0kai.stone.annotations.component.Component
 import com.github.klee0kai.stone.annotations.component.ExtendOf
-import com.github.klee0kai.stone.annotations.component.Init
-import com.github.klee0kai.stone.annotations.component.ModuleOriginFactory
 import com.github.klee0kai.thekey.app.BuildConfig
 import com.github.klee0kai.thekey.app.di.debug.DebugDI
 import com.github.klee0kai.thekey.app.di.debug.DebugDI.initDummyModules
 import com.github.klee0kai.thekey.app.di.dependencies.AppComponentProviders
-import com.github.klee0kai.thekey.app.di.modules.AndroidHelpersModule
 import com.github.klee0kai.thekey.app.di.modules.CoreAndroidHelpersModuleFactory
-import com.github.klee0kai.thekey.app.di.modules.CoroutineModule
-import com.github.klee0kai.thekey.app.di.modules.DBModule
-import com.github.klee0kai.thekey.app.di.modules.EngineModule
-import com.github.klee0kai.thekey.app.di.modules.HelpersModule
-import com.github.klee0kai.thekey.app.di.modules.InteractorsModule
-import com.github.klee0kai.thekey.app.di.modules.PresentersModule
-import com.github.klee0kai.thekey.app.di.modules.RepositoriesModule
 import com.github.klee0kai.thekey.app.features.allFeatures
 import com.github.klee0kai.thekey.app.features.model.findApi
 import com.github.klee0kai.thekey.app.ui.navigation.deeplink.configMainDeeplinks
@@ -50,59 +40,10 @@ var DI: AppComponent = initAppComponent()
         AppWrappersStone::class,
     ],
 )
-interface AppComponent : CoreComponent, AppComponentProviders {
+interface AppComponent : AppComponentProviders, AppComponentModules, CoreComponent {
 
     @ExtendOf
     fun ext(component: CoreComponent)
-
-
-    /* get module */
-    fun coroutine(): CoroutineModule
-    fun presenters(): PresentersModule
-    fun androidHelpers(): AndroidHelpersModule
-    fun helpers(): HelpersModule
-    fun interactors(): InteractorsModule
-    fun repositories(): RepositoriesModule
-    fun engine(): EngineModule
-    fun databases(): DBModule
-
-    /* get origin factories */
-    @ModuleOriginFactory
-    fun coroutineFactory(): CoroutineModule
-
-    @ModuleOriginFactory
-    fun presentersFactory(): PresentersModule
-
-    @ModuleOriginFactory
-    fun androidHelpersFactory(): AndroidHelpersModule
-
-    @ModuleOriginFactory
-    fun helpersFactory(): HelpersModule
-
-    @ModuleOriginFactory
-    fun interactorsFactory(): InteractorsModule
-
-    @ModuleOriginFactory
-    fun repositoriesFactory(): RepositoriesModule
-
-    @ModuleOriginFactory
-    fun engineFactory(): EngineModule
-
-    @ModuleOriginFactory
-    fun databasesFactory(): DBModule
-
-    /* override */
-    @Init
-    fun initEngineModule(engineModule: Class<out EngineModule>)
-
-    @Init
-    fun initHelpersModule(helpers: Class<out HelpersModule>)
-
-    @Init
-    fun initAndroidHelpersModule(helpers: AndroidHelpersModule)
-
-    @Init
-    fun initPresenterModule(presentersModule: PresentersModule)
 
 }
