+++ app_mobile/src/main/kotlin/com/github/klee0kai/thekey/app/di/dependencies/AppComponentProviders.kt
@@ -4,7 +4,6 @@ interface AppComponentProviders :
     PresentersDependencies,
     AndroidHelpersDependencies,
     HelperDependencies,
-    CoroutineDependencies,
     InteractorsDependencies,
     RepositoriesDependencies,
     EngineDependencies,
