+++ core/src/main/kotlin/com/github/klee0kai/thekey/core/di/CoreComponent.kt
@@ -6,17 +6,13 @@ import com.github.klee0kai.stone.KotlinWrappersStone
 import com.github.klee0kai.stone.Stone
 import com.github.klee0kai.stone.annotations.component.Component
 import com.github.klee0kai.stone.annotations.component.GcWeakScope
-import com.github.klee0kai.stone.annotations.component.Init
-import com.github.klee0kai.stone.annotations.component.ModuleOriginFactory
 import com.github.klee0kai.stone.annotations.component.RunGc
 import com.github.klee0kai.stone.annotations.module.BindInstance
+import com.github.klee0kai.thekey.core.di.dependecies.CoreDependencyProvider
 import com.github.klee0kai.thekey.core.di.identifiers.NoteGroupIdentifier
 import com.github.klee0kai.thekey.core.di.identifiers.NoteIdentifier
 import com.github.klee0kai.thekey.core.di.identifiers.PluginIdentifier
 import com.github.klee0kai.thekey.core.di.identifiers.StorageIdentifier
-import com.github.klee0kai.thekey.core.di.modules.CoreAndroidHelpersModule
-import com.github.klee0kai.thekey.core.di.modules.CoreInteractorsModule
-import com.github.klee0kai.thekey.core.di.modules.ThemeModule
 import com.github.klee0kai.thekey.core.di.wrap.AppWrappersStone
 import com.github.klee0kai.thekey.core.domain.model.AppConfig
 import com.github.klee0kai.thekey.core.domain.model.feature.model.DynamicFeature
@@ -39,7 +35,7 @@ var CoreDI: CoreComponent = initCoreComponent()
         AppWrappersStone::class,
     ],
 )
-interface CoreComponent : CoreDependencyProvider {
+interface CoreComponent : CoreDependencyProvider, CoreComponentModules {
 
     @BindInstance(cache = BindInstance.CacheType.Weak)
     fun ctx(ctx: Context? = null): Context
@@ -53,33 +49,6 @@ interface CoreComponent : CoreDependencyProvider {
     @RunGc
     @GcWeakScope
     fun gcWeak()
-
-
-    /* get module */
-    fun theme(): ThemeModule
-    fun coreAndroidHelpersModule(): CoreAndroidHelpersModule
-    fun coreInteractors(): CoreInteractorsModule
-
-    /* get origin factories */
-    @ModuleOriginFactory
-    fun themeFactory(): ThemeModule
-
-    @ModuleOriginFactory
-    fun coreAndroidHelpersModuleFactory(): CoreAndroidHelpersModule
-
-    @ModuleOriginFactory
-    fun coreInteractorsFactory(): CoreInteractorsModule
-
-
-    /* set origin factories */
-    @Init
-    fun initThemeModule(themeModule: ThemeModule)
-
-    @Init
-    fun initCoreAndroidHelpersModule(themeModule: CoreAndroidHelpersModule)
-
-    @Init
-    fun initCoreInteractorsModule(interactorsModule: CoreInteractorsModule)
 
 }
