--- core/src/main/kotlin/com/github/klee0kai/thekey/core/di/CoreDependencyProvider.kt
@@ -1,8 +1,0 @@
-package com.github.klee0kai.thekey.core.di
-
-import com.github.klee0kai.thekey.core.di.dependecies.CoreAndroidHelpersDependencies
-import com.github.klee0kai.thekey.core.di.dependecies.CoreInteractorsDependencies
-
-interface CoreDependencyProvider :
-    CoreAndroidHelpersDependencies,
-    CoreInteractorsDependencies
