+++ core/src/main/kotlin/com/github/klee0kai/thekey/core/di/dependecies/CoreDependencyProvider.kt
@@ -1,0 +1,20 @@
+package com.github.klee0kai.thekey.core.di.dependecies
+
+import android.content.Context
+import androidx.activity.ComponentActivity
+import com.github.klee0kai.stone.annotations.dependencies.Dependencies
+import com.github.klee0kai.thekey.core.domain.model.AppConfig
+
+@Dependencies
+interface CoreDependencyProvider :
+    CoreAndroidHelpersDependencies,
+    CoreInteractorsDependencies,
+    CoroutineDependencies {
+
+    fun ctx(): Context
+
+    fun activity(): ComponentActivity?
+
+    fun config(): AppConfig
+
+}
