+++ core/src/main/kotlin/com/github/klee0kai/thekey/core/di/modules/CoroutineModule.kt
@@ -1,4 +1,4 @@
-package com.github.klee0kai.thekey.app.di.modules
+package com.github.klee0kai.thekey.core.di.modules
 
 import com.github.klee0kai.stone.annotations.module.Module
 import com.github.klee0kai.stone.annotations.module.Provide
