+++ libs.versions.toml
@@ -28,7 +28,7 @@ ml-vision = "1.4.0-alpha05"
 jetbrain-immutable = "0.3.7"
 jetbrain-coroutines = "1.8.0"
 
-stone = "1.0.6"
+stone = "1.0.7"
 hummus = "0.0.2"
 
 room = "2.6.1"
